requirements.txt
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
Authlib==1.3.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
fastembed==0.2.5
filelock==3.13.4
flatbuffers==24.3.25
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.62.1
grpcio-health-checking==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
numpy==1.26.4
onnx==1.16.0
onnxruntime==1.17.3
openai==1.14.3
packaging==24.0
pandas==2.2.2
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qdrant-client==1.8.2
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
validators==0.22.0
wcwidth==0.2.13
weaviate-client==4.5.4
zipp==3.18.1
